requirements.txt
Django
djangorestframework
django-cors-headers
pika
psycopg2
psycopg2-binary